fix(kit): `InputMonth` hydration
