+++ projects/kit/components/input-month/index.ts
@@ -1,4 +1,5 @@
 export * from './input-month';
 export * from './input-month.component';
+export * from './input-month-content.component';
 export * from './input-month.directive';
 export * from './input-month.options';
