--- projects/kit/components/input-month/input-month.template.html
@@ -1,12 +1,0 @@
-@if (host.native) {
-    <input
-        *tuiTextfieldContent
-        type="month"
-        [max]="max()?.toJSON()"
-        [min]="min()?.toJSON()"
-        [value]="host.value()?.toJSON()"
-        (click.stop.zoneless)="(0)"
-        (input)="onInput($any($event.target).value)"
-        (pointerdown.stop.zoneless)="(0)"
-    />
-}
